ダイアログフォント決定方法見直し#2
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/branches/cmake@7380 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/common/dlglib_tmpl.cpp
@@ -392,7 +392,7 @@ static DLGTEMPLATE *GetDlgTemplate(
 		new_size = prev_size;
 	} else {
 		// copy with replacing font
-		size_t size_namediff =
+		int size_namediff =
 			wcslen(FontFaceName) - wcslen(logfont.lfFaceName);
 		size_namediff *= sizeof(WCHAR);
 		size_namediff += 3;		// テキスト、フォント名長でalignが変化する
